feature: move normal enemy modules into normal namespace.
