+++ src/in_game/enemy.rs
@@ -1,25 +1,22 @@
 mod assets_holder;
 mod barrage;
-mod emerge;
 mod life_count;
-mod movement;
+mod normal;
 mod system_label;
 
 use crate::app_state::AppState;
 pub use barrage::bullet::Bullet;
 use barrage::EnemyBarragePlugin;
 use bevy::prelude::*;
-use emerge::EnemyEmergePlugin;
 use life_count::EnemyLifeCountPlugin;
-use movement::EnemyMovementPlugin;
 
 pub struct EnemyPlugin;
 
 impl Plugin for EnemyPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugin(EnemyBarragePlugin)
-            .add_plugin(EnemyEmergePlugin)
-            .add_plugin(EnemyMovementPlugin)
+            .add_plugin(normal::NormalEnemyEmergePlugin)
+            .add_plugin(normal::NormalEnemyMovementPlugin)
             .add_plugin(EnemyLifeCountPlugin)
             .add_system_set(SystemSet::on_enter(AppState::InGame).with_system(setup))
             .add_system_set(SystemSet::on_exit(AppState::InGame).with_system(cleanup));
