+++ src/in_game/enemy/normal.rs
@@ -1,0 +1,5 @@
+mod emerge;
+mod movement;
+
+pub use emerge::NormalEnemyEmergePlugin;
+pub use movement::NormalEnemyMovementPlugin;
