+++ src/in_game/enemy/normal/movement/move_pattern.rs
