+++ src/in_game/enemy/normal/movement/move_pattern/down_stay_right_bottom.rs
