+++ src/in_game/enemy/normal/movement/move_pattern/down_stay_up.rs
