+++ src/in_game/enemy/normal/movement/move_pattern/fast_down_right.rs
