+++ src/in_game/enemy/normal/movement/move_pattern/left_bottom.rs
