+++ src/in_game/enemy/normal/movement/move_pattern/right_bottom.rs
