+++ src/in_game/enemy/system_label.rs
@@ -2,5 +2,4 @@ use bevy::prelude::*;
 #[derive(Debug, Clone, PartialEq, Eq, Hash, SystemLabel)]
 pub enum EnemySystemLabel {
     LifeCount,
-    EmergeCount,
 }
